+++ bugsnag-android-core/src/main/java/com/bugsnag/android/InternalReportDelegate.java
@@ -97,15 +97,14 @@ class InternalReportDelegate implements EventStore.Delegate {
         event.addMetadata(INTERNAL_DIAGNOSTICS_TAB, "notifierVersion", notifier.getVersion());
         event.addMetadata(INTERNAL_DIAGNOSTICS_TAB, "apiKey", immutableConfig.getApiKey());
 
-        final EventPayload eventPayload = new EventPayload(null, event, notifier);
+        final EventPayload payload = new EventPayload(null, event, notifier);
         try {
             Async.run(new Runnable() {
                 @Override
                 public void run() {
                     try {
                         Delivery delivery = immutableConfig.getDelivery();
-                        String apiKey = eventPayload.getApiKey();
-                        DeliveryParams params = immutableConfig.getErrorApiDeliveryParams(apiKey);
+                        DeliveryParams params = immutableConfig.getErrorApiDeliveryParams(payload);
 
                         // can only modify headers if DefaultDelivery is in use
                         if (delivery instanceof DefaultDelivery) {
@@ -113,7 +112,7 @@ class InternalReportDelegate implements EventStore.Delegate {
                             headers.put(HEADER_INTERNAL_ERROR, "true");
                             headers.remove(DeliveryHeadersKt.HEADER_API_KEY);
                             DefaultDelivery defaultDelivery = (DefaultDelivery) delivery;
-                            defaultDelivery.deliver(params.getEndpoint(), eventPayload, headers);
+                            defaultDelivery.deliver(params.getEndpoint(), payload, headers);
                         }
 
                     } catch (Exception exception) {